+++ nix/example-cluster.nix
@@ -3,7 +3,7 @@
     enable = true;
     package = ./result;
     nodeName = "node-a@10.0.0.11";
-    cookieFile = ./cluster.cookie;
+    cookieFile = "/etc/nixos/nix-swarm/secrets/swarm.cookie";
 
     peers = [
       "node-a@10.0.0.11"
@@ -21,6 +21,7 @@
       replicas = 2;
       unitTemplate = "gitea@%{slot}.service";
       constraints = [ "ssd" ];
+      preferredNodes = [ "node-a@10.0.0.11" "node-b@10.0.0.12" ];
       healthcheck = "/run/current-system/sw/bin/curl -fsS http://127.0.0.1:3000/";
       settings = {
         domain = "gitea.home";